requirements_minimal.txt
# please make sure you have already a pytorch install that is cuda enabled!
dora-search
diffq>=0.2.1
einops
julius>=0.2.3
lameenc>=1.2
openunmix
pyyaml
torch>=1.8.1
torchaudio>=0.8
tqdm

# above requirements are for demucs
# below requirements are for video generation
pandas
plotly